requirements.txt
Pillow
requests
transformers==4.30.2
timm
sentencepiece
accelerate==0.20.3